Move & rename function_not_called
