+++ tests/coroutines_asyncio/test_context_propagation.py
@@ -15,7 +15,7 @@
 import sys
 
 import pytest
-from testing_support.fixtures import function_not_called, override_generic_settings
+from testing_support.validators.validate_function_not_called import validate_function_not_called, override_generic_settings
 from testing_support.validators.validate_transaction_metrics import validate_transaction_metrics
 
 from newrelic.api.application import application_instance as application
@@ -126,7 +126,7 @@ def test_context_propagation(event_loop, schedule, loop_policy):
 
 
 @override_generic_settings(global_settings(), {"enabled": False})
-@function_not_called("newrelic.core.stats_engine", "StatsEngine.record_transaction")
+@validate_function_not_called("newrelic.core.stats_engine", "StatsEngine.record_transaction")
 def test_nr_disabled(event_loop):
     import asyncio
